Fixing unit tests.
