+++ evalml/automl/engine/__init__.py
@@ -1,3 +1,3 @@
-from .engine_base import EngineBase, EngineComputation, train_pipeline, train_and_score_pipeline
+from .engine_base import EngineBase, EngineComputation, train_pipeline, train_and_score_pipeline, evaluate_pipeline
 from .sequential_engine import SequentialEngine
 from .dask_engine import DaskEngine
